Refactor code - minimize import statements using init module
